Extract result.error logic to error handler, include context in exception extension
